Resolved conflicting code due to package refactoring from upstream.
